chore: move AxesTuple out of _internal
