+++ src/boost_histogram/_internal/hist.py
@@ -27,9 +27,8 @@ import numpy as np
 import boost_histogram
 from boost_histogram import _core
 
+from ..axis import AxesTuple, Axis, Variable
 from ..storage import Double, Storage
-from .axestuple import AxesTuple
-from .axis import Axis, Variable
 from .enum import Kind
 from .typing import Accumulator, ArrayLike, CppHistogram
 from .utils import cast, register, set_module
